changed to money values to decimal(15,2)

+++ src/Aggregator.DTOs/AggregatorMoneyDto.cs
@@ -8,7 +8,7 @@ public class AggregatorMoneyDto
     /// <summary>
     /// Сумма в минимальных единицах валюты
     /// </summary>
-    public long Amount { get; set; }
+    public decimal Amount { get; set; }
     
     /// <summary>
     /// Трехзначный числовой код валюты (ISO-4217)
